Moved view model sources to their own folder
