+++ src/main/java/uk/ac/cam/groupseven/weatherapp/viewmodelsources/HomeWeatherSource.java
@@ -1,7 +1,9 @@
-package uk.ac.cam.groupseven.weatherapp.datasources;
+package uk.ac.cam.groupseven.weatherapp.viewmodelsources;
 
 import com.google.inject.Inject;
 import io.reactivex.Observable;
+import uk.ac.cam.groupseven.weatherapp.datasources.RowingInfoSource;
+import uk.ac.cam.groupseven.weatherapp.datasources.WeatherSource;
 import uk.ac.cam.groupseven.weatherapp.models.FlagStatus;
 import uk.ac.cam.groupseven.weatherapp.models.Weather;
 import uk.ac.cam.groupseven.weatherapp.viewmodels.HomeWeather;
