+++ src/main/java/uk/ac/cam/groupseven/weatherapp/viewmodelsources/HoursWeatherSource.java
@@ -1,7 +1,8 @@
-package uk.ac.cam.groupseven.weatherapp.datasources;
+package uk.ac.cam.groupseven.weatherapp.viewmodelsources;
 
 import com.google.inject.Inject;
 import io.reactivex.Observable;
+import uk.ac.cam.groupseven.weatherapp.datasources.OpenWeatherSource;
 import uk.ac.cam.groupseven.weatherapp.models.Weather;
 import uk.ac.cam.groupseven.weatherapp.viewmodels.HourWeather;
