+++ src/test/java/uk/ac/cam/groupseven/weatherapp/mocking/MockedModule.java
@@ -4,13 +4,17 @@ import com.google.inject.*;
 import com.google.inject.name.Names;
 import uk.ac.cam.groupseven.weatherapp.Screen;
 import uk.ac.cam.groupseven.weatherapp.ScreenLayout;
-import uk.ac.cam.groupseven.weatherapp.datasources.*;
-import uk.ac.cam.groupseven.weatherapp.viewmodels.CrestViewModel;
+import uk.ac.cam.groupseven.weatherapp.datasources.CrestSource;
+import uk.ac.cam.groupseven.weatherapp.datasources.RowingInfoSource;
+import uk.ac.cam.groupseven.weatherapp.datasources.WeatherSource;
 import uk.ac.cam.groupseven.weatherapp.models.FlagStatus;
 import uk.ac.cam.groupseven.weatherapp.models.Weather;
 import uk.ac.cam.groupseven.weatherapp.models.Wind;
+import uk.ac.cam.groupseven.weatherapp.viewmodels.CrestViewModel;
 import uk.ac.cam.groupseven.weatherapp.viewmodels.HomeWeather;
 import uk.ac.cam.groupseven.weatherapp.viewmodels.HourWeather;
+import uk.ac.cam.groupseven.weatherapp.viewmodelsources.CrestViewModelSource;
+import uk.ac.cam.groupseven.weatherapp.viewmodelsources.ViewModelSource;
 
 import java.awt.*;
 import java.nio.file.Path;
