+++ src/test/java/uk/ac/cam/groupseven/weatherapp/mocking/ViewModelSourceMock.java
@@ -2,7 +2,7 @@ package uk.ac.cam.groupseven.weatherapp.mocking;
 
 import com.google.inject.Inject;
 import io.reactivex.Observable;
-import uk.ac.cam.groupseven.weatherapp.datasources.ViewModelSource;
+import uk.ac.cam.groupseven.weatherapp.viewmodelsources.ViewModelSource;
 
 public class ViewModelSourceMock<T> implements ViewModelSource<T> {
     private T mock;
